Move some code and added create group after user signup
