+++ controllers/localLoginController.js
@@ -1,8 +1,11 @@
 const passport = require('passport')
 const express = require('express')
 const User = require('../models/user')
+const Group = require('../models/Group')
 const router = express.Router()
 router.use(express.static("public"))
+const LocalStrategy = require('passport-local').Strategy
+passport.use(new LocalStrategy(User.authenticate()));
 
 router.post('/signup', async (req, res, next) => {
   await User.register(new User({ username: req.body.username }), req.body.password, (err) => {
@@ -17,14 +20,20 @@ router.post('/signup', async (req, res, next) => {
           res.status(400)
           return res.send({ message: err })
         } else {
-          req.logIn(user, (err) => {
+          req.logIn(user, async (err) => {
             if (err) { return next(err) }
+            const createNewGRP = await Group.create({
+              grpName: `${req.body.username} personal-group`,
+              members: [user._id],
+              ownerID: user._id
+            })
+            await User.findByIdAndUpdate(user._id, { groups: createNewGRP._id }, { new: true })
             return res.send({ message: "User registered and login Successful" })
           })
         }
       })(req, res, next)
     }
-  });
+  })
 })
 
 router.post('/login', async (req, res, next) => {
@@ -50,25 +59,23 @@ router.get('/getlogin', (req, res) => {
 })
 
 router.get('/checkusername/:username', async (req, res) => {
-  const findUser =  await User.findOne({ "username": req.params.username }) 
-    if (!findUser) {
-      return res.send({ message: "user not found" })
-    }
-    
-    return res.send({ message: "user found" })
-  })
+  const findUser = await User.findOne({ "username": req.params.username })
+  if (!findUser) {
+    return res.send({ message: "user not found" })
+  }
+
+  return res.send({ message: "user found" })
+})
 
 router.post('/setnewpassword', async (req, res) => {
   const foundUser = await User.findOne({ "username": req.body.username })
   await foundUser.setPassword(req.body.password);
   const updatePassword = await foundUser.save();
- if(updatePassword)
- {
-    if (!updatePassword)
-    {
-      res.send({message: "Password update failure (Server error)"})
+  if (updatePassword) {
+    if (!updatePassword) {
+      res.send({ message: "Password update failure (Server error)" })
     }
-    return res.send({message: "Password updated"})
+    return res.send({ message: "Password updated" })
   }
 })
