+++ package.json
@@ -13,6 +13,8 @@
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "express-session": "^1.17.2",
+    "http-status-codes": "^2.2.0",
+    "lodash": "^4.17.21",
     "method-override": "^3.0.0",
     "mongoose": "^6.1.2",
     "nodemon": "^2.0.15",
